requirements.txt
torch>=1.13.1
transformers>=4.28.0
numpy>=1.21.0
requests>=2.28.0
nltk>=3.8.1
tqdm>=4.65.0
rank_bm25>=0.2.2
scikit-learn>=1.0.2

# Optional, highly recommended for dense retrieval acceleration
faiss-cpu>=1.7.3  # or faiss-gpu depending on your environment
